# apiLayer/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Api UMbook",
        "description": "Aqui encontra-se a documentação de toda a api de dados da UMbook, incluindo todas as rotas e toda a informação à cerca do que é precisar passar como parâmetros, bem como os resultados que se podem esperar ao executar aquela rota.\n Muitas das rotas encontram-se protegidas, não se esqueça de realizar o login na aba dos utilizadores com as suas credenciais para ter acesso às rotas que lhe são permitido aceder. Receberá um token de acesso e terá que o copiar e colar no botão 'Authorize' no canto superior direito."
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:3049",
    "basePath": "/api",
...
    "paths": {
        "/cursos": {
            "get": {
                "tags":[
                    "cursos"
                ],
                "summary": "Todos os ids dos cursos registados.",
                "description": "Retorna uma lista com todos os ids dos cursos presentes e respetivas designações.",
                "responses": {
                    "200": {
                        "description": "Lista de Curso",
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "curso": {
                                        "type": "string"
                                    },
                                    "designacao": {
                                      "type":"string"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post":{
                "tags":[
                    "cursos"
                ],
                "summary": "Inserir um novo curso.",
                "description": "Adiciona um novo curso à base de dados.",
                "parameters": [
                    {
                      "name": "id",
                      "in": "formData",
                      "description": "ID do curso",
                      "required": true,
                      "type": "string"
                    },
                    {
                      "name": "nome",
                      "in": "formData",
                      "description": "Nome do curso",
                      "required": true,
                      "type": "string"
                    }
                  ],
                "responses": {
                    "200": {
                        "description": "Retorna o id do Curso criado",
                        "schema": {
                            "type": "string",
                            "items": {
                                "properties": {
                                    "curso": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "401":{
                      "description": "Falta-lhe o token de acesso ou o token não é válido."
                    }
                },
                "security": [
                  {
                    "api_key": []
                  }
                ]
            }
        },
        "/cursos/{cursoId}": {
            "get": {
              "tags": [
                "cursos"
              ],
              "summary": "Dado um identificador de um curso, devolve toda a informação relativa a esse curso.",
              "description": "Devolve toda a informação relativa a um curso.",
              "operationId": "getCursoById",
              "parameters": [
                {
                  "name": "cursoId",
                  "in": "path",
                  "description": "ID do curso a retornar.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                    "description": "Retorna toda a informação relativa a um curso.",
                    "schema": {
                        "type": "object",
                        "properties": {
                            "info": {
                                "$ref": "#/definitions/Curso"
                            },
                            "anos":{
                                "type":"array",
                                "items": {
                                    "type":"object",
                                    "properties": {
                                        "id":{
                                            "type":"string"
                                        },
                                        "designacao":{
                                            "type":"string"
                                        }
                                    }
                                }
                            },
                            "estudantes":{
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/Utilizador"
                                }
                            },
                            "responsaveis":{
                                "type": "object",
                                "properties": {
                                    "ano":{
                                        "type":"string"
                                    },
                                    "responsaveisAno":{
                                        "type":"array",
                                        "items":{
                                            "$ref": "#/definitions/Utilizador"
                                        }
                                    }
                                }
                            },
                            "publicacoes":{
                                "type":"array",
                                "items": {
                                    "$ref": "#/definitions/Publicacao"
                                }
                            }
                        }
                    }
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                },
                "403":{
                  "description": "Não possuí nível de acesso suficiente."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
            },
            "put":{
              "tags": [
                "cursos"
              ],
              "summary": "Dado um identificador de um curso, edite a informação relacionada com ele.",
              "description": "Para editar um curso, é necessário ser Admin.",
              "operationId": "putCursoById",
              "parameters": [
                {
                  "name": "cursoId",
                  "in": "path",
                  "description": "ID do curso a editar.",
                  "required": true,
                  "type": "string"
                },
                {
                  "name": "designacao",
                  "in": "formData",
                  "description": "Nova designação.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                  "description": "O curso foi editado."
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                },
                "403":{
                  "description": "Não possuí nível de acesso suficiente."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
            },
            "delete": {
              "tags": [
                "cursos"
              ],
              "summary": "Dado um identificador de um curso, elimine o curso da base de dados.",
              "description": "Para apagar um curso, é necessário ser Admin.",
              "operationId": "deleteCursoById",
              "parameters": [
                {
                  "name": "cursoId",
                  "in": "path",
                  "description": "ID do curso a editar.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                  "description": "O curso foi eliminado."
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                },
                "403":{
                  "description": "Não possuí nível de acesso suficiente."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
            }
        },
        "/cursos/{cursoId}/anos": {
            "get": {
              "tags": [
                "cursos"
              ],
              "summary": "Dado um identificador de um curso, devolve todos os anos pertencentes a esse curso.",
              "description": "Devolve todos os anos de um curso.",
              "operationId": "getAnosCursoById",
              "parameters": [
                {
                  "name": "cursoId",
                  "in": "path",
                  "description": "ID do curso a retornar.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                    "description": "Retorna todos os anos relativos a um curso.",
                    "schema": {
                        "type": "array",
                        "items": {
                            "type":"object",
                            "properties": {
                                "id":{
                                    "type":"string"
                                },
                                "designacao":{
                                    "type":"string"
                                }
                            }
                        }
                    }
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                }
              }
            }
        },
        "/cursos/{cursoId}/estudantes": {
            "get": {
              "tags": [
                "cursos"
              ],
              "summary": "Dado um identificador de um curso, devolve todos os estudantes pertencentes a esse curso.",
              "description": "Devolve todos os estudantes de um curso.",
              "operationId": "getEstudantesAnoById",
              "parameters": [
                {
                  "name": "cursoId",
                  "in": "path",
                  "description": "ID do curso a retornar.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                    "description": "Retorna todos os estudantes relativos a um curso.",
                    "schema": {
                        "type":"array",
                        "items": {
                           "$ref": "#/definitions/Utilizador"
                        }
                    }
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
            }
        },
        "/cursos/{cursoId}/publicacoes": {
            "get": {
              "tags": [
                "cursos"
              ],
              "summary": "Dado um identificador de um curso, devolve todas as publicações pertencentes a esse curso.",
              "description": "Devolve todos as publicações de um curso.",
              "operationId": "getPublicacoesCursoById",
              "parameters": [
                {
                  "name": "cursoId",
                  "in": "path",
                  "description": "ID do curso a retornar.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                    "description": "Retorna todos as publicações relativos a um curso.",
                    "schema": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/Publicacao"
                        }
                    }
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
            }
        },
        "/cursos/{cursoId}/responsaveis": {
            "get": {
              "tags": [
                "cursos"
              ],
              "summary": "Dado um identificador de um curso, devolve todas os responsáveis pertencentes a esse curso.",
              "description": "Devolve todos os responsáveis de um curso.",
              "operationId": "getResponsaveisCursoById",
              "parameters": [
                {
                  "name": "cursoId",
                  "in": "path",
                  "description": "ID do curso a retornar.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                    "description": "Retorna todos os responsáveis relativos a um curso.",
                    "schema": {
                        "type":"array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "ano":{
                                    "type":"string"
                                },
                                "responsaveisAno":{
                                    "type":"array",
                                    "items":{
                                        "$ref": "#/definitions/Utilizador"
                                    }
                                }
                            }
                        }
                    }
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
            }
        },
        "/anos": {
            "post":{
                "tags":[
                    "anos"
                ],
                "summary": "Inserir um novo ano de um dado curso.",
                "description": "Para inserir um novo ano a um curso, é necessário ser Admin.",
                "parameters": [
                    {
                      "name": "nome",
                      "in": "formData",
                      "description": "Nome do ano",
                      "required": true,
                      "type": "string"
                    },
                    {
                        "name": "anoLetivo",
                        "in": "formData",
                        "description": "Ano letivo do ano",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "idCurso",
                        "in": "formData",
                        "description": "ID do curso que o ano pertence",
                        "required": true,
                        "type": "string"
                      }
                  ],
                "responses": {
                    "200": {
                        "description": "Retorna o id do ano criado",
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401":{
                      "description": "Falta-lhe o token de acesso ou o token não é válido."
                    },
                    "403":{
                      "description": "Não possuí nível de acesso suficiente."
                    }
                },
                "security": [
                  {
                    "api_key": []
                  }
                ]
            }
        },
        "/anos/{anoId}": {
            "get": {
              "tags": [
                "anos"
              ],
              "summary": "Dado um identificador de um ano, devolve toda a informação relativa a esse ano.",
              "description": "Devolve toda a informação relativa a um ano.",
              "operationId": "getAnoById",
              "parameters": [
                {
                  "name": "anoId",
                  "in": "path",
                  "description": "ID do ano a retornar.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                    "description": "Retorna toda a informação relativa a um ano.",
                    "schema": {
                        "type": "object",
                        "properties": {
                            "info": {
                                "$ref": "#/definitions/Ano"
                            },
                            "cadeiras":{
                                "type":"array",
                                "items": {
                                    "type":"object",
                                    "properties": {
                                        "id":{
                                            "type":"string"
                                        },
                                        "designacao":{
                                            "type":"string"
                                        }
                                    }
                                }
                            },
                            "estudantes":{
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/Utilizador"
                                }
                            },
                            "responsaveis":{
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/Utilizador"
                                }
                            },
                            "publicacoes":{
                                "type":"array",
                                "items": {
                                    "$ref": "#/definitions/Publicacao"
                                }
                            }
                        }
                    }
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
            },
            "put":{
              "tags":[
                "anos"
            ],
            "summary": "Editar um determinado ano.",
            "description": "Para editar um ano, é necessário ser Admin ou Responsável desse ano.",
            "parameters": [
                {
                  "name": "anoId",
                  "in": "path",
                  "description": "ID do ano a retornar.",
                  "required": true,
                  "type": "string"
                },
                {
                  "name": "nome",
                  "in": "formData",
                  "description": "Nome do ano",
                  "required": true,
                  "type": "string"
                },
                {
                    "name": "anoLetivo",
                    "in": "formData",
                    "description": "Ano letivo do ano",
                    "required": true,
                    "type": "string"
                }
              ],
            "responses": {
                "200": {
                    "description": "O ano foi editado com sucesso."
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                },
                "403":{
                  "description": "Não possuí nível de acesso suficiente."
                }
            },
            "security": [
              {
                "api_key": []
              }
            ]
            },
            "delete": {
              "tags":[
                "anos"
            ],
            "summary": "Eliminar um determinado ano.",
            "description": "Para apagar um ano, é necessário ser Admin.",
            "parameters": [
                {
                  "name": "anoId",
                  "in": "path",
                  "description": "ID do ano a eliminar.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                    "description": "O ano foi apagado com sucesso."
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                },
                "403":{
                  "description": "Não possuí nível de acesso suficiente."
                }
            },
            "security": [
              {
                "api_key": []
              }
            ]
            }
        },
...
        "/anos/{anoId}/publicacoes": {
            "get": {
              "tags": [
                "anos"
              ],
              "summary": "Dado um identificador de um ano, devolve todas as publicações pertencentes a esse ano.",
              "description": "Devolve todos as publicações de um ano.",
              "operationId": "getPublicacoesAnoById",
              "parameters": [
                {
                  "name": "anoId",
                  "in": "path",
                  "description": "ID do curso a retornar.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                    "description": "Retorna todas as publicações relativos a um curso.",
                    "schema": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/Publicacao"
                        }
                    }
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
            }
        },
        "/anos/{anoId}/responsaveis": {
            "get": {
              "tags": [
                "anos"
              ],
              "summary": "Dado um identificador de um ano, devolve todas os responsáveis pertencentes a esse ano.",
              "description": "Devolve todos os responsáveis de um ano.",
              "operationId": "getResponsaveisAnoById",
              "parameters": [
                {
                  "name": "anoId",
                  "in": "path",
                  "description": "ID do ano a retornar.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                    "description": "Retorna todos os responsáveis relativos a um ano.",
                    "schema": {
                        "type":"array",
                        "items": {
                           "$ref": "#/definitions/Utilizador"
                        }
                    }
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
            }
        },
        "/anos/{anoId}/estudantes": {
            "get": {
              "tags": [
                "anos"
              ],
              "summary": "Dado um identificador de um ano, devolve todas os estudantes pertencentes a esse ano.",
              "description": "Devolve todos os estudantes de um ano.",
              "operationId": "getEstudantesAnoById",
              "parameters": [
                {
                  "name": "anoId",
                  "in": "path",
                  "description": "ID do ano a retornar.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                    "description": "Retorna todos os estudantes relativos a um ano.",
                    "schema": {
                        "type":"array",
                        "items": {
                           "$ref": "#/definitions/Utilizador"
                        }
                    }
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
            }
        },
...
        "/publicacoes":{
            "post":{
                "tags":[
                    "publicacoes"
                ],
                "summary": "Inserir uma nova publicação.",
                "description": "Adiciona uma nova publicação à base de dados.",
                "parameters": [
                    {
                      "name": "idGrupo",
                      "in": "formData",
                      "description": "Id do grupo em que a publicação vai ser publicada",
                      "required": true,
                      "type": "string"
                    },
                    {
                        "name": "conteudo",
                        "in": "formData",
                        "description": "Conteudo da publicação",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "idUtilizador",
                        "in": "formData",
                        "description": "Id do utilizador que a publicou",
                        "required": true,
                        "type": "string"
                    }
                  ],
                "responses": {
                    "200": {
                        "description": "Retorna o id da publicação criada",
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401":{
                      "description": "Falta-lhe o token de acesso ou o token não é válido."
                    }
                },
                "security": [
                  {
                    "api_key": []
                  }
                ]
            }
        },
...
    },
      "securityDefinitions":{
        "api_key":{
          "type": "apiKey",
          "in": "query",
          "name": "token"
        }
      },
      "security": [{
         "api_key": []
      }
    ],
    "definitions": {
        "Utilizador": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "nomeUtilizador": {
                "type": "string"
              },
              "numeroAluno": {
                "type": "string"
              },
              "dataNascimento": {
                "type": "string"
              },
              "numeroTelemovel": {
                "type": "string"
              },
              "sexo": {
                "type": "string"
              }
            }
        },
        "Curso":{
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "designacao": {
                "type": "string"
              }
            }
        },
...
        "Publicacao":{
            "type": "object",
            "properties": {
                "info":{
                    "type":"object",
                    "properties": {
                        "id":{
                            "type":"string"
                        },
                        "conteudo":{
                            "type":"string"
                        },
                        "data":{
                            "type":"string"
                        },
                        "likes":{
                            "type":"integer"
                        },
                        "idUtilizador":{
                            "type":"string"
                        }
                    }
                },
                "ficheiros":{
                    "type":"array",
                    "items": {
                        "$ref": "#/definitions/Ficheiro"
                    }
                },
                "comentarios":{
                    "type":"array",
                    "items": {
                        "$ref": "#/definitions/Comentario"
                    }
                }
            }
        },
...
    }   
}